feat(speed): 优化测速精度，支持小数点速度值
- 将节点速度字段从int改为float64，提升测速精度
- 调整速度显示格式，保留两位小数
- 优化最小速度配置步长，支持0.1MB/s调节
- 完善Mihomo测速逻辑，改进超时处理机制
- 更新相关API和日志输出，支持浮点数速度值

+++ models/node.go
@@ -17,7 +17,7 @@ type Node struct {
 	Source          string `gorm:"default:'manual'"`
 	SourceID        int
 	Group           string
-	Speed           int    `gorm:"default:0"` // 测速结果(MB/s)
+	Speed           float64 `gorm:"default:0"` // 测速结果(MB/s)
 	DelayTime       int    `gorm:"default:0"` // 延迟时间(ms)
 	LastCheck       string // 最后检测时间
 }
